requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile 'E:\mic\requirements.in'
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
attrs==25.4.0
    # via sarif-om
blis==1.3.3
    # via thinc
cachetools==6.2.2
    # via google-auth
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   nltk
    #   typer-slim
    #   uvicorn
cloudpathlib==0.23.0
    # via weasel
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   wasabi
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
diffusers==0.35.2
    # via -r E:/mic/requirements.in
diskcache==5.6.3
    # via llama-cpp-python
fastapi==0.121.3
    # via -r E:/mic/requirements.in
filelock==3.20.0
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.10.0
    # via
    #   huggingface-hub
    #   torch
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via
    #   -r E:/mic/requirements.in
    #   google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r E:/mic/requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
googletrans==4.0.2
    # via -r E:/mic/requirements.in
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via -r E:/mic/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx[http2]==0.28.1
    # via
    #   -r E:/mic/requirements.in
    #   googletrans
huggingface-hub==0.36.0
    # via
    #   diffusers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via diffusers
jinja2==3.1.6
    # via
    #   llama-cpp-python
    #   spacy
    #   torch
joblib==1.5.2
    # via nltk
llama-cpp-python==0.3.16
    # via -r E:/mic/requirements.in
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
murmurhash==1.0.15
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.5
    # via torch
nltk==3.9.2
    # via -r E:/mic/requirements.in
numpy==2.2.6
    # via
    #   blis
    #   diffusers
    #   llama-cpp-python
    #   opencv-python
    #   spacy
    #   thinc
    #   transformers
opencv-python==4.12.0.88
    # via -r E:/mic/requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   huggingface-hub
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pbr==7.0.3
    # via sarif-om
pillow==11.3.0
    # via diffusers
preshed==3.0.12
    # via
    #   spacy
    #   thinc
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psutil==7.1.3
    # via -r E:/mic/requirements.in
psycopg2-binary==2.9.11
    # via -r E:/mic/requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.4
    # via
    #   confection
    #   fastapi
    #   google-generativeai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.2.1
    # via -r E:/mic/requirements.in
python-json-logger==4.0.0
    # via -r E:/mic/requirements.in
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via
    #   diffusers
    #   nltk
    #   transformers
requests==2.32.5
    # via
    #   diffusers
    #   google-api-core
    #   huggingface-hub
    #   spacy
    #   transformers
    #   weasel
rsa==4.9.1
    # via google-auth
safetensors==0.7.0
    # via
    #   diffusers
    #   transformers
sarif-om==1.0.4
    # via -r E:/mic/requirements.in
scapy==2.6.1
    # via -r E:/mic/requirements.in
smart-open==7.5.0
    # via weasel
sniffio==1.3.1
    # via anyio
spacy==3.8.11
    # via -r E:/mic/requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.50.0
    # via fastapi
sympy==1.14.0
    # via torch
thinc==8.3.10
    # via spacy
tokenizers==0.22.1
    # via transformers
torch==2.9.1
    # via -r E:/mic/requirements.in
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   nltk
    #   spacy
    #   transformers
transformers==4.57.1
    # via -r E:/mic/requirements.in
typer-slim==0.20.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   huggingface-hub
    #   llama-cpp-python
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r E:/mic/requirements.in
waitress==2.1.2
    # via -r E:/mic/requirements.in
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.3
    # via spacy
wrapt==2.0.1
    # via smart-open
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
gunicorn 
